+++ pints/tests/test_opt_nelder_mead.py
@@ -68,8 +68,8 @@ class TestNelderMead(unittest.TestCase):
         # Best position and score called before run
         self.assertEqual(list(opt.x_best()), list(x))
         self.assertEqual(list(opt.x_guessed()), list(x))
-        self.assertEqual(opt.f_best(), float('inf'))
-        self.assertEqual(opt.f_guessed(), float('inf'))
+        self.assertEqual(opt.f_best(), np.inf)
+        self.assertEqual(opt.f_guessed(), np.inf)
 
         # Not running
         self.assertFalse(opt.running())